Listing updated

+++ src/app/modules/listings/listing.service.ts
@@ -2,8 +2,11 @@ import { prisma } from "../../shared/prisma";
 import { ICreateListingPayload } from "./listing.interface";
 import { AppError } from "../../errors/AppError";
 import httpStatus from "http-status";
-import { v2 as cloudinary } from "cloudinary";
 import { JwtPayload } from "jsonwebtoken";
+import calculatePagination, { IOptions } from "../../helpers/paginationHelper";
+import { FilterParams } from "../../../constants";
+import { Prisma } from "@prisma/client";
+import { listingSearchableFields } from "./listing.constant";
 
 const createListing = async (token: JwtPayload, payload: ICreateListingPayload) => {
 
@@ -24,10 +27,60 @@ const createListing = async (token: JwtPayload, payload: ICreateListingPayload) 
     return listing;
 };
 
-const getAllListings = async () => {
-    return await prisma.listing.findMany({
+const getAllListings = async (params: FilterParams, options: IOptions) => {
+    const { page, limit, skip, sortBy, sortOrder } = calculatePagination(options)
+    const { searchTerm, ...filterData } = params;
+
+    const andConditions: Prisma.ListingWhereInput[] = [];
+    if (searchTerm) {
+        andConditions.push({
+            OR: listingSearchableFields.map(field => ({
+                [field]: {
+                    contains: searchTerm,
+                    mode: "insensitive"
+                }
+            }))
+        })
+    };
+
+    // if (specialties && specialties.length > 0) {
+    //     andConditions.push({
+    //         doctorSpecialties: {
+    //             some: {
+    //                 specialities: {
+    //                     title: {
+    //                         contains: specialties,
+    //                         mode: "insensitive"
+    //                     }
+    //                 }
+    //             }
+    //         }
+    //     })
+    // };
+
+    if (Object.keys(filterData).length > 0) {
+        andConditions.push({
+            AND: Object.keys(filterData).map(key => ({
+                [key]: {
+                    equals: (filterData as any)[key]
+                }
+            }))
+        })
+    };
+
+    const whereConditions: Prisma.ListingWhereInput = andConditions.length > 0 ? {
+        AND: andConditions
+    } : {};
+
+    const result = await prisma.listing.findMany({
+        skip,
+        take: limit,
+        where: whereConditions,
+        orderBy: { [sortBy]: sortOrder },
         include: { guide: true, bookings: true, reviews: true }
     });
+
+    return result;
 };
 
 const getSingleListing = async (token: JwtPayload, id: string) => {
@@ -75,7 +128,25 @@ const updateListing = async (token: JwtPayload, id: string, payload: Partial<ICr
     });
 };
 
-const deleteListing = async (id: string) => {
+const deleteListing = async (token: JwtPayload, id: string) => {
+    const isExistUser = await prisma.user.findUnique({
+        where: {
+            email: token.email
+        }
+    });
+
+    if (!isExistUser) {
+        throw new AppError(httpStatus.BAD_REQUEST, "User not found");
+    };
+
+    const isExistListing = await prisma.listing.findUnique({
+        where: { id }
+    });
+
+    if (!isExistListing) {
+        throw new AppError(httpStatus.BAD_REQUEST, "Listing not found");
+    };
+
     return await prisma.listing.delete({
         where: { id }
     });
